129 - Move data creation module out of data_pipelines
