+++ tests/tel2puml/legacy_loop_detection/walk_puml_graph/test_node_legacy.py
@@ -24,7 +24,7 @@ from tel2puml.events import (
 from tel2puml.legacy_loop_detection.events import (
     get_event_reference_from_events,
 )
-from tel2puml.data_pipelines.data_creation import (
+from tel2puml.data_creation import (
     generate_test_data_event_sequences_from_puml,
 )
